requirements.txt
aiohttp
aiomysql
python-dateutil
langdetect
PyYAML
feedparser
beautifulsoup4
python-dotenv
openai